+++ CHANGELOG.rst
@@ -23,10 +23,12 @@ Added
   flattened instance with elements drawn randomly from the original instance.
 - Add transpose method to all 3D classes to transpose navigation dimensions.
 - Reading of a crystal map from orientation data in Bruker's HDF5 file format.
+- Uniform sampling of orientation space using cubochoric sampling.
 
 Changed
 -------
-- to_euler() changed internally, "Krakow_Hielscher" deprecated, use "MTEX" instead
+- to_euler() changed internally, "Krakow_Hielscher" deprecated, use "MTEX" instead.
+- Default orientation space sampling method from "haar_euler" to "cubochoric".
 
 2021-05-23 - version 0.6.0
 ==========================